+++ src/nmea_converter/m_conv_nmea_char_course.cpp
@@ -42,7 +42,7 @@ int conv_NMEA(const char* message, float* latitude, float* longitude, float* t_c
 				size = comma_f - comma_l;
 				float_conv = (char*)calloc(size, sizeof(char));
 				//strncpy(float_conv, &message[comma_l + 1], size - 1);
-				strncpy_s(float_conv, size, &message[comma_l + 1], size - 1);
+				strncpy_s(float_conv, size, &message[comma_l + 1], (size_t)(size)-1ull);
 				*latitude = atof(float_conv);
 
 				degrees = floor(*latitude / 100);
@@ -65,7 +65,7 @@ int conv_NMEA(const char* message, float* latitude, float* longitude, float* t_c
 				size = comma_f - comma_l;
 				float_conv = (char*)calloc(size, sizeof(char));
 				//strncpy(float_conv, &message[comma_l + 1], size - 1);
-				strncpy_s(float_conv, size, &message[comma_l + 1], size - 1);
+				strncpy_s(float_conv, size, &message[comma_l + 1], (size_t)(size)-1ull);
 
 				*longitude = atof(float_conv);
 
@@ -87,7 +87,7 @@ int conv_NMEA(const char* message, float* latitude, float* longitude, float* t_c
 				case 9://True course
 					size = comma_f - comma_l;
 					float_conv = (char*)calloc(size, sizeof(char));
-					strncpy_s(float_conv, size, &message[comma_l + 1], size - 1);
+					strncpy_s(float_conv, size, &message[comma_l + 1], (size_t)(size)-1ull);
 
 					*t_course = atof(float_conv);
 
@@ -100,7 +100,7 @@ int conv_NMEA(const char* message, float* latitude, float* longitude, float* t_c
 
 					size = comma_f - comma_l;
 					float_conv = (char*)calloc(size, sizeof(char));
-					strncpy_s(float_conv, size, &message[comma_l + 1], size - 1);
+					strncpy_s(float_conv, size, &message[comma_l + 1], (size_t)(size)-1ull);
 
 					variation = atof(float_conv);
